some partial work on getting registrations working
It has been sometime since this code has been touched :( Just checking it all in with the hopes that it not in to bad of shape.  And if it is I'll just have to work through it.  Yeah not idea but I just want to pick this project up again after a pretty long hiatus.